Fix #8958: Add initialization ordering and remove user/bot initialization from Authorizer
