+++ openmetadata-service/src/main/java/org/openmetadata/service/jdbi3/UserRepository.java
@@ -27,6 +27,7 @@ import java.util.stream.Collectors;
 import javax.ws.rs.core.UriInfo;
 import lombok.extern.slf4j.Slf4j;
 import org.openmetadata.schema.api.teams.CreateTeam.TeamType;
+import org.openmetadata.schema.auth.SSOAuthMechanism;
 import org.openmetadata.schema.entity.teams.AuthenticationMechanism;
 import org.openmetadata.schema.entity.teams.Team;
 import org.openmetadata.schema.entity.teams.User;
@@ -34,15 +35,18 @@ import org.openmetadata.schema.type.EntityReference;
 import org.openmetadata.schema.type.Include;
 import org.openmetadata.schema.type.Relationship;
 import org.openmetadata.service.Entity;
+import org.openmetadata.service.OpenMetadataApplicationConfig;
 import org.openmetadata.service.exception.CatalogExceptionMessage;
 import org.openmetadata.service.jdbi3.CollectionDAO.EntityRelationshipRecord;
 import org.openmetadata.service.resources.teams.UserResource;
 import org.openmetadata.service.secrets.SecretsManager;
 import org.openmetadata.service.secrets.SecretsManagerFactory;
+import org.openmetadata.service.security.SecurityUtil;
 import org.openmetadata.service.security.policyevaluator.SubjectCache;
 import org.openmetadata.service.util.EntityUtil;
 import org.openmetadata.service.util.EntityUtil.Fields;
 import org.openmetadata.service.util.JsonUtils;
+import org.openmetadata.service.util.UserUtil;
 
 @Slf4j
 public class UserRepository extends EntityRepository<User> {
@@ -179,6 +183,22 @@ public class UserRepository extends EntityRepository<User> {
 
   public boolean checkEmailAlreadyExists(String emailId) {
     return daoCollection.userDAO().checkEmailExists(emailId) > 0;
+  }
+
+  public void initializeUsers(OpenMetadataApplicationConfig config) {
+    Set<String> adminUsers = new HashSet<>(config.getAuthorizerConfiguration().getAdminPrincipals());
+    LOG.debug("Checking user entries for admin users {}", adminUsers);
+    String domain = SecurityUtil.getDomain(config);
+    String providerType = config.getAuthenticationConfiguration().getProvider();
+    if (providerType.equals(SSOAuthMechanism.SsoServiceType.BASIC.value())) {
+      UserUtil.handleBasicAuth(adminUsers, domain);
+    } else {
+      UserUtil.addUsers(adminUsers, domain, true);
+    }
+
+    LOG.debug("Checking user entries for test users");
+    Set<String> testUsers = new HashSet<>(config.getAuthorizerConfiguration().getTestPrincipals());
+    UserUtil.addUsers(testUsers, domain, null);
   }
 
   private List<EntityReference> getOwns(User user) throws IOException {
