+++ openmetadata-service/src/main/java/org/openmetadata/service/resources/policies/PolicyResource.java
@@ -102,6 +102,7 @@ public class PolicyResource extends EntityResource<Policy, PolicyRepository> {
     // Load any existing rules from database, before loading seed data.
     dao.initSeedDataFromResources();
     ResourceRegistry.initialize(listOrEmpty(PolicyResource.getResourceDescriptors()));
+    PolicyCache.initialize();
   }
 
   @Override
